+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file. This projec
 to [Semantic Versioning] (http://semver.org/). For change log format,
 use [Keep a Changelog] (http://keepachangelog.com/).
 
+## [Unreleased]
+
+### Changed
+
+- Move factory methods for uploaded files to `Factory\UploadedFileFactoryTrait`
+
+### Deprecated
+
+- `UploadedFile::parseUploadedFiles`, use `Factory\UploadedFileFactoryTrait::createUploadedFiles()` instead
+
 ## [2.5.0] - 2025-03-14
 
 ### Changed